Refactor export_utilities into depth_charts
